+++ src/Layout/PageContainer.js
@@ -1,21 +1,21 @@
-import { Box, CssBaseline, Drawer, IconButton, List, ListItem, ListItemButton, ListItemIcon, ListItemText, ThemeProvider, Typography } from "@mui/material"
+import { Box, CssBaseline, IconButton, ThemeProvider, useMediaQuery, useTheme } from "@mui/material"
 import MenuIcon from '@mui/icons-material/Menu';
 import DarkModeIcon from '@mui/icons-material/DarkMode';
 import LightModeIcon from '@mui/icons-material/LightMode';
 import { useState } from "react";
 import { darkTheme, lightTheme } from "../theme";
 import { Outlet } from "react-router-dom";
-import { navbarItems } from "../pages/home/utils";
+import { useDrawer } from "../drawerContext";
 
 const PageContainer = () => {
-    const [drawer, setDrawer] = useState(false)
+    const muiTheme = useTheme()
+    const isSmallScreen = useMediaQuery(muiTheme.breakpoints.down('sm'))
+    const { toggleDrawer } = useDrawer()
+
     const [theme, setTheme] = useState(darkTheme)
 
-    const toggleDrawer = () => setDrawer(!drawer)
     const handleSelectDarkTheme = () => setTheme(darkTheme)
     const handleSelectLightTheme = () => setTheme(lightTheme)
-
-    
 
     return (
         <ThemeProvider theme={theme}>
@@ -30,22 +30,11 @@ const PageContainer = () => {
                         <DarkModeIcon />
                     </IconButton>
                 )}
-                <IconButton onClick={toggleDrawer} sx={{ m: 1 }}>
-                    <MenuIcon />
-                </IconButton>
-                
-                <Drawer anchor='right' open={drawer} onClose={toggleDrawer}>
-                    <List sx={{ minWidth: 300 }}>
-                        {navbarItems.map(item => (
-                            <ListItem disablePadding key={item.text}>
-                                <ListItemButton>
-                                    <ListItemIcon><Typography fontSize='large'>{item.icon}</Typography></ListItemIcon>
-                                    <ListItemText primary={<Typography color={`${item.color}.main`}>{item.text}</Typography>} />
-                                </ListItemButton>
-                            </ListItem>
-                        ))}
-                    </List>
-                </Drawer>
+                {isSmallScreen && (
+                    <IconButton onClick={toggleDrawer} sx={{ m: 1 }}>
+                        <MenuIcon />
+                    </IconButton>
+                )}
             </Box>
             <main>
                 <Outlet />
